requirements/common.txt
Flask==3.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Bootstrap4==4.0.2
WTForms==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.35
Werkzeug==3.0.4
wheel==0.44.0
Flask-Login==0.6.3
Flask-Moment==1.0.5
markdown==3.6